cwm: fix USB mounting
this lun= config is needed because recovery currently uses an old
kernel. the recovery partition is small and an old kernel is used to
let it all fit in. build variables could be used too.

an alternative fix would be to find a .config that generated a smaller
new kernel and to use that instead of the old recovery_kernel. then
(I think) the lun should be on a standard path that cwm already
understands.